edclineage-bulk-export/props/requirements.txt
requests==2.27.1
urllib3==1.26.9
pandas==1.4.1
numpy==1.22.3
lxml==4.8.0
python-dotenv==0.20.0
bs4==0.0.1
psycopg2==2.9.3